Split into kernel crate
