+++ crates/collections/src/lib.rs
@@ -1,6 +1,4 @@
-pub use self::{array::*, seq_buf::*, seq_iter::*, vec2d::*};
-
-mod array;
-mod seq_buf;
-mod seq_iter;
-mod vec2d;
+pub mod array;
+pub mod seq_buf;
+pub mod seq_iter;
+pub mod vec2d;
