+++ crates/kernel/README.md
@@ -1,0 +1,3 @@
+<!-- cargo-sync-rdme title [[ -->
+# elden-analyzer-kernel
+<!-- cargo-sync-rdme ]] -->
