+++ crates/kernel/src/types/mod.rs
@@ -1,0 +1,3 @@
+pub mod clip_rect;
+pub mod rect;
+pub mod time;
