+++ src/bin/elden_analyzer/subcommand/analyze/comp_accum.rs
@@ -3,9 +3,10 @@ use std::{collections::VecDeque, sync::mpsc};
 use color_eyre::eyre;
 use elden_analyzer::{
     components::{ComponentContainer, Detection, DetectionPayload},
-    video_capture::{Frame, FramePosition},
+    video_capture::Frame,
 };
-use elden_analyzer_collections::SeqIter;
+use elden_analyzer_collections::seq_iter::SeqIter;
+use elden_analyzer_kernel::types::time::FramePosition;
 
 use super::comp_detect;
