+++ src/bin/elden_analyzer/subcommand/analyze/comp_detect.rs
@@ -3,8 +3,9 @@ use std::time::Instant;
 use color_eyre::eyre;
 use elden_analyzer::{
     components::{Component, ComponentContainer, Components, Detection},
-    video_capture::{Frame, FramePosition},
+    video_capture::Frame,
 };
+use elden_analyzer_kernel::types::time::FramePosition;
 
 use super::decode;
