--- src/geometry/mod.rs
@@ -1,4 +1,0 @@
-pub use self::{clip::*, neighbor::*};
-
-mod clip;
-mod neighbor;
