--- src/image_process/ui_finder.rs
@@ -1,1 +1,0 @@
-
