+++ src/lib.rs
@@ -1,6 +1,5 @@
 pub mod algorithm;
 pub mod components;
-pub mod geometry;
 pub mod image_process;
 pub mod operator;
 pub mod traits;
